added Navbar and about page
